Modified compareMatch.cpp and hotBrain.py
I have a basic example working for this program. It is able to process each file and displays the output at the end in a similar format to our database structure.

hotBrain.py now includes an example of how we can run c++ executables with python. Commented out functions are still in progress and not meant to do anything just yet.