Created dataFile server communication flow
Updated the backend to now process all videos for the user and create the respective files. Then the program will run the algorithm to find the matches and process that into a CSV. Now the file is able to be sent to the server.

Had to modify the algorithm file to account for redundancies.

Also modified the server side endpoint to read the data in correctly.