Matches now display only user preferences
Created the algorithm to display the matches based on both user and match preferences.

It will parse out any matches that do not fit the preferences for the current user or the match user.
It will sort in ascending order from lowest to highest score (lower being better match).
It will filter out matches from the list to only display the top 10.

Also updated the git ignore to ignore .csv files and updated the backend to remove the extra datafile created for matches.